[Diopucr-Manu] * Integración con Portafirmas.

+++ tramitador/ispac-catalog/ispac-catalog-java/src/main/java/ieci/tdw/ispac/ispaccatalog/action/sustitucion/GestionSustitucionesAction.java
@@ -4,6 +4,7 @@ import ieci.tdw.ispac.api.IInvesflowAPI;
 import ieci.tdw.ispac.api.IRespManagerAPI;
 import ieci.tdw.ispac.api.ISecurityAPI;
 import ieci.tdw.ispac.api.errors.ISPACException;
+import ieci.tdw.ispac.api.errors.ISPACRuleException;
 import ieci.tdw.ispac.api.impl.SessionAPI;
 import ieci.tdw.ispac.api.item.IItem;
 import ieci.tdw.ispac.api.item.IItemCollection;
@@ -15,14 +16,24 @@ import ieci.tdw.ispac.ispaclib.bean.BeanFormatter;
 import ieci.tdw.ispac.ispaclib.bean.CacheFormatterFactory;
 import ieci.tdw.ispac.ispaclib.bean.CollectionBean;
 import ieci.tdw.ispac.ispaclib.bean.ItemBean;
+import ieci.tdw.ispac.ispaclib.dao.security.SustitucionDAO;
 import ieci.tdw.ispac.ispaclib.resp.Responsible;
+import ieci.tdw.ispac.ispaclib.sign.portafirmas.ProcessSignConnector;
+import ieci.tdw.ispac.ispaclib.sign.portafirmas.ProcessSignConnectorFactory;
 import ieci.tdw.ispac.ispaclib.utils.DateUtil;
 import ieci.tdw.ispac.ispaclib.utils.StringUtils;
+import ieci.tecdoc.sgm.core.exception.SigemException;
+import ieci.tecdoc.sgm.core.services.LocalizadorServicios;
+import ieci.tecdoc.sgm.core.services.entidades.Entidad;
+import ieci.tecdoc.sgm.core.services.estructura_organizativa.EstructuraOrganizativaException;
+import ieci.tecdoc.sgm.core.services.estructura_organizativa.ServicioEstructuraOrganizativa;
+import ieci.tecdoc.sgm.core.services.estructura_organizativa.Usuario;
 
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -32,6 +43,8 @@ import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
 import org.apache.struts.action.ActionMessage;
 import org.apache.struts.action.ActionMessages;
+
+import es.dipucr.sigem.api.rule.common.utils.EntidadesAdmUtil;
 
 public class GestionSustitucionesAction extends BaseDispatchAction {
 	
@@ -77,24 +90,52 @@ public class GestionSustitucionesAction extends BaseDispatchAction {
 				String descripcion = frm.getProperty("DESCRIPCION");
 				String uidSustituido = frm.getProperty("UID_SUSTITUIDO");
 				String sustituteIds = frm.getProperty("UIDS_SUSTITUTOS");
-				if ((StringUtils.isNotEmpty(uidSustituido)) &&
-					(StringUtils.isNotEmpty(sustituteIds))) {
-					
-					String[] uidSustitutos = sustituteIds.split(",");
-					if (uidSustitutos.length > 0) {
-						
-						// Generar las sustituciones para el período establecido
-						security.addSustituciones(uidSustituido, uidSustitutos, fechaInicio, fechaFin, descripcion);
-					}
+				
+				
+				if ((StringUtils.isNotEmpty(uidSustituido)) && (StringUtils.isNotEmpty(sustituteIds))) {
+					if(uidSustituido.equals(sustituteIds)){
+						errors.add(ActionMessages.GLOBAL_MESSAGE, new ActionMessage("No pueden ser iguales el sustituto/sustituido"));
+					}else{
+						String[] uidSustitutos = sustituteIds.split(",");
+						if (uidSustitutos.length > 0) {						
+							for (String sustituto : uidSustitutos) {								
+								insertarAutorizacion(session, sustituto, uidSustituido, fechaInicio, fechaFin, descripcion);
+							}
+							// Generar las sustituciones para el período establecido
+							security.addSustituciones(uidSustituido, uidSustitutos, fechaInicio, fechaFin, descripcion);
+						}
+					}					
 				}
 				else {
 					// Modificación del período de sustitución
 					String idFechSustitucion = frm.getProperty("ID_FECH_SUSTITUCION");
 					if (StringUtils.isNotEmpty(idFechSustitucion)) {
+						// Revocamos la antigua sustitucion
+						IItemCollection fm_responsibles = SustitucionDAO.getSustitucion(session.getClientContext().getConnection(), Integer.parseInt(idFechSustitucion)).disconnect();
+						while (fm_responsibles.next()) {				
+							IItem itemSubstitute = fm_responsibles.value();
+							String uid_sustituido = itemSubstitute.getString("SUSTITUCION:UID_SUSTITUIDO");
+							String uid_sustituto = itemSubstitute.getString("SUSTITUCION:UID_SUSTITUTO");
+							Date fecha_inicioAntigua = itemSubstitute.getDate("FECHASUST:FECHA_INICIO");
+							Date fecha_finAntigua = itemSubstitute.getDate("FECHASUST:FECHA_FIN");
+							
+							if(uidSustituido.equals(sustituteIds)){
+								errors.add(ActionMessages.GLOBAL_MESSAGE, new ActionMessage("No pueden ser iguales el sustituto/sustituido"));
+							}else{
+								revocamosAutorizacion(session, uid_sustituido, uid_sustituto, fecha_inicioAntigua, fecha_finAntigua);
+								
+								//Insertamos con las nuevas fechas
+								insertarAutorizacion(session, uid_sustituto, uid_sustituido, fechaInicio, fechaFin, descripcion);
 						
-						security.modifyFechSustitucion(Integer.parseInt(idFechSustitucion), fechaInicio, fechaFin, descripcion);
+							}
+						}
+						if (errors.isEmpty()) {
+							security.modifyFechSustitucion(Integer.parseInt(idFechSustitucion), fechaInicio, fechaFin, descripcion);
+						}
+						
 					}
 				}
+				
 				
 				return mapping.findForward("closeFrame");
 	        }
@@ -104,6 +145,71 @@ public class GestionSustitucionesAction extends BaseDispatchAction {
     	return mapping.findForward("success_viewAddFecha");
 	}
 	
+	private void insertarAutorizacion(SessionAPI session, String sustituto, String uidSustituido, Date fechaInicio, Date fechaFin, String descripcion) throws ISPACRuleException {
+		try{
+			String [] idUsuario = uidSustituido.split("-");
+			String [] idUsuarioSustituto = sustituto.split("-");
+			ServicioEstructuraOrganizativa servicioEO = LocalizadorServicios.getServicioEstructuraOrganizativa();		
+			Entidad entidad = EntidadesAdmUtil.obtenerEntidadObject(session.getClientContext());		
+			if(idUsuario.length==2 && idUsuarioSustituto.length==2){
+				if(idUsuario[0].equals("1") && idUsuarioSustituto[0].equals("1")){							
+					Usuario usuarioAutoriza = servicioEO.getUsuario(Integer.parseInt(idUsuario[1]), entidad.getIdentificador());
+					Usuario usuarioAutorizado = servicioEO.getUsuario(Integer.parseInt(idUsuarioSustituto[1]), entidad.getIdentificador());	            	
+					ProcessSignConnector portafirmasSignConnector = ProcessSignConnectorFactory.getInstance(EntidadesAdmUtil.obtenerEntidadObject(session.getClientContext()).getIdentificador()).getProcessSignConnector();
+					portafirmasSignConnector.insertaAutorizacionPortafirmas(fechaInicio, fechaFin,  usuarioAutoriza.getUserData().getDni(), usuarioAutorizado.getUserData().getDni(), descripcion, EntidadesAdmUtil.obtenerEntidad(session.getClientContext()));
+				}
+	
+			}
+		} catch (NumberFormatException e) {
+			logger.error("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+		} catch (ISPACException e) {
+			logger.error("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+		} catch (EstructuraOrganizativaException e) {
+			logger.error("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+		} catch (SigemException e) {
+			logger.error("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido insertarAutorizacion. " + e.getMessage(), e);
+		} 
+		
+	}
+
+	private void revocamosAutorizacion(SessionAPI session, String uid_sustituido, String uid_sustituto, Date fecha_inicio, Date fecha_fin) throws ISPACException {
+		try {
+			ServicioEstructuraOrganizativa servicioEO = LocalizadorServicios.getServicioEstructuraOrganizativa();
+			Entidad entidad = EntidadesAdmUtil.obtenerEntidadObject(session.getClientContext());			
+
+			if (uid_sustituido != null && fecha_inicio != null && fecha_fin != null && uid_sustituto != null) {
+				String[] idUsuario = uid_sustituido.split("-");
+				String[] idUsuarioSustituto = uid_sustituto.split("-");
+				if (idUsuario.length == 2 && idUsuarioSustituto.length == 2) {
+					if (idUsuario[0].equals("1") && idUsuarioSustituto[0].equals("1")) {
+						Usuario usuarioAutoriza = servicioEO.getUsuario(Integer.parseInt(idUsuario[1]), entidad.getIdentificador());
+						Usuario usuarioAutorizado = servicioEO.getUsuario(Integer.parseInt(idUsuarioSustituto[1]), entidad.getIdentificador());
+						ProcessSignConnector portafirmasSignConnector = ProcessSignConnectorFactory.getInstance(
+								EntidadesAdmUtil.obtenerEntidadObject(session.getClientContext()).getIdentificador()).getProcessSignConnector();
+						portafirmasSignConnector.revocarAutorizacion(fecha_inicio, fecha_fin, usuarioAutoriza.getUserData().getDni(), usuarioAutorizado.getUserData().getDni(), EntidadesAdmUtil.obtenerEntidad(session.getClientContext()));
+					}
+				}
+			}
+
+		} catch (NumberFormatException e) {
+			logger.error("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+		} catch (EstructuraOrganizativaException e) {
+			logger.error("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+		} catch (SigemException e) {
+			logger.error("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+		} catch (ISPACException e) {
+			logger.error("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+			throw new ISPACRuleException("No se ha podido eliminar deleteSubstitute. " + e.getMessage(), e);
+		}
+	}
+
 	public ActionForward viewFechaSustitucion(ActionMapping mapping, 
 			 								  ActionForm form,
 			 								  HttpServletRequest request, 
@@ -493,14 +599,28 @@ public class GestionSustitucionesAction extends BaseDispatchAction {
 		
 		SustitutionsForm deleteSupervisedsForm = (SustitutionsForm) form;
 		String[] multibox = deleteSupervisedsForm.getMultibox();
-		
+		boolean eliminado = false;
 		if (multibox != null && multibox.length > 0 ) {
 			
 			ISecurityAPI securityAPI = session.getAPI().getSecurityAPI();
 			
-			securityAPI.deleteSustituciones(multibox);
+			for (int i = 0; i < multibox.length; i++) {			
+
+				Map<String, Object> resultado = securityAPI.getSustitucion(multibox[i]);
+				String uid_sustituido = (String) resultado.get("uid_sustituido");
+				String uid_sustituto = (String) resultado.get("uid_sustituto");
+				Date fecha_inicio = (Date)resultado.get("fecha_inicio");
+				Date fecha_fin = (Date)resultado.get("fecha_fin");
+				if(!uid_sustituido.equals(uid_sustituto)){
+					revocamosAutorizacion(session, uid_sustituido, uid_sustituto, fecha_inicio, fecha_fin);
+					eliminado=true;
+				}				
+			}			
+
+			if(eliminado){
+				securityAPI.deleteSustituciones(multibox);
+			}
 		}
-		
 	}
 	
 	public ActionForward deleteSubstituteHistoric(ActionMapping mapping,
